Fix Array and add function jsGet
Now when the item is an array, it will be written to a map as an array and inside it will have its pure data.

And added new jsGet function to read object in oriented way.

each value now has its proper data, sequence to string,
atom for numbers, integer for boolean.